+++ rust/healpix-geo-core/src/index/mod.rs
@@ -1,2 +1,2 @@
-mod index;
 mod indexers;
+mod region;
