+++ rust/healpix-geo-core/src/index/region.rs
